filter: Move all pytests into test_filter
As I mentioned in Slack [1], I think tests should reflect the
organization of the files being tested.
Since everything is under filter.py right now, all pytests for that file
should live in test_filter.py. This is consistent with having all Cram
tests for that file under tests/functional/filter/cram/.

[1]: https://bedfordlab.slack.com/archives/C01LCTT7JNN/p1653072798818139?thread_ts=1653072787.820709&cid=C01LCTT7JNN
